requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
graphrag==2.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
jinja2>=3.1.2
pyarrow>=15.0.0,<16.0.0
openai>=1.57.0,<2.0.0
aiofiles>=24.1.0,<25.0.0